glabal progress

+++ src/main/scala/fr/rebaze/domain/services/MetricsService.scala
@@ -8,7 +8,7 @@ import java.time.LocalDate
 trait MetricsService:
   def getMetricsByDay(day: LocalDate): Task[Seq[Metric]]
   val extractRulesIdFromJsonDirectExport: Task[Seq[String]]
-  def getGlobalProgressByUserId(userId: String): Task[Float]
+  def getGlobalProgressByUserId(userId: String): Task[Double]
 object MetricsService:
   def getMetricsByDay(day: LocalDate): RIO[MetricsService, Seq[Metric]] =
     ZIO.serviceWithZIO[MetricsService](_.getMetricsByDay(day))
@@ -16,5 +16,5 @@ object MetricsService:
   val extractRulesIdFromJsonDirectExport: RIO[MetricsService, Seq[String]] =
     ZIO.serviceWithZIO[MetricsService](_.extractRulesIdFromJsonDirectExport)
 
-  def getGlobalProgressByUserId(userId: String): RIO[MetricsService, Float] =
+  def getGlobalProgressByUserId(userId: String): RIO[MetricsService, Double] =
     ZIO.serviceWithZIO[MetricsService](_.getGlobalProgressByUserId(userId))
